Update hovered entity via MouseMoved events
Doing it this way instead of getting the mouse position in the update
function means that GUI layers can consume these inputs first so
entities will not get lit up when the mouse is over a window.

* Also implemented the concept of consuming events without blocking
them. This way, a MouseMovedEvent can be consumed by ImGui but it can
still get sent to the Selector system to deselect entities.