Move all camera logic to Lua scripts (#2)
The issue with the current setup is that movement logic was implemented via different types of camera derived off of a base camera interface. This is obviously strange, and the PlayerBasedCamera suggests a way forward; designating an entity to be the camera, attaching a camera component that stores the projection matrix, and then calculating the view matrix from the entities position. The latter part is what the StaticCamera did.

For the movement in FirstPersonCamera and ThirdPersonCamera, we can shift this to Lua scripts, fully separating the movement logic from the camera logic; in the new way, it is simply an entity that contains a camera component and a script component.

### TL;DR
* Added CameraComponent.
* Gave every entity a separate LuaEngine.
* Remove Camera and all derived classes.
* Added Lua scripts for First and Third person cameras.

### Notable Commits
* 6ddfff9320f5bfc91ed7257912b274a5a7a99724 - FirstPersonCamera in Lua
* 04027d71fb3d8c0e6539a051b33a213d1c617d70 - Switch to using an entity with a Lua script instead of a first person camera
* b12ce74ca62fc462bb1e08720a6b43c69ea39fe3 - Give each Entity a separate LuaEngine
* 19ac20845d00f41d4ca079e82438db7af9a8743c - Move LuaEngines inside the ScriptRunner
* 5139f19ae787ceeda7265e1b5dd6ed4bde1ea113 - Got 3rd person movement working in Lua