Simplify PhysicsComponent (#16)
PhysicsComponent used to store a variant of different collider types, which added a slight bit of complexity which isn't really needed. I have also been looking into serialisation and having to serialise a variant seemed needlessly complicated.

The main reason the variant existed was because different colliders have different attributes. But since there are only a few for now and I have no current plans to enhance it just yet, I've removed the variant and just put all the different attribute types in the struct. This does lead to some redundant data in the structs, for instance all sphere colliders have a halfExtents vector now, but that's only 4 floats and we don't really have a shortage of space. This will make physics component easier to serialise.