DevUI Cleanup (#34)
DevUI is the Sprocket wrapper for ImGui, however it was mainly just bloat. This removes a huge chunk of it, with the intention being to just use the ImGui types and functions directly rather than having it all wrapped by Sprocket. In the future, we should add a SPROCKET_INCLUDE_IMGUI compile flag so that Sprocket can be used without ImGui if needed.