Scene Pausing (#43)
Previously, we would "pause" a scene by just not calling the update function. This creates some issues, in particular if a scene system makes use of the mouse. If a scene does not get updated, then when we unpause, the mouse offset could be huge, which was causing the camera to jump when closing the console in `Runtime`. 

To fix this, we do a similar thing to what we did with events. Previously, objects would "consume" events just by not passing them on, but this also caused issues with mouse and keyboard events for disabled/paused/inactive objects. We changed events to have an `consumed` flag. Then objects would still be made aware of consumed events.

Now, Scene::OnUpdate and all system update functions receive an `active` flag, which is true if the scene is running and false if it is paused. It is up to the systems themselves to decide what to do when paused. Now, they can continue to update their internal mouse proxies without needing to update the whole system.

This change also adds some comments here and there and some code cleaning, as well as making it possible to change the time of day in `Game` from the escape menu and have it be visible without closing the menu.