Remove Entity Sorting (#46)
Regularly sorting the elements to optimise the rendering was poor design, as sorting can be quite inefficient, and made the renderer more complex. Instead, we can just build up a map of draw commands each frame to achieve the same with a minimal performance hit. This allows us to remove the sorting feature. The main motivation for this was so that we do not need to implement sorting in our own ECS, which will greatly simplify the design.

- Change the renderer to build up a map of draw commands. This no longer relies on the entities being sorted. This also means all frames are done in the fewest draw commands, whereas before sorting would only occur once every 5 seconds, causing new entities to be separate draw calls until then.
- Remove the sorting in `Scene`.
- Fix the initial position of the camera in `Anvil`.
- Extract material uploading into a separate function to reduce duplication.