Adding systems can now be done after adding entities
When calling Registry::OnAdd<T>, the given callback will be invoked for all existing
entities with component T. This was done because copying a scene resets
the target scene, which wiped all systems. Wiping the scene is necessary
since a copy overwrites the internal structures of the registry so that
entity IDs can be preserved.

This does introduce a possible nasty bug if we are resetting a
previously used scene and storing handles to entities; it is possible
that a handle does not get invalidated since the ID could be active
again after the copy and assumptions have been made on the components of
that entity, however this should never happen. Really, if you are
copying a scene, you should be copying into a fresh scene anyway. To
enforce this, it may be worthwhile changing the signature of the
Loader::Copy function to return the new scene instead, but how should it
get returned? There are use cases for value type, shared_ptr and
unique_ptr, maybe support all three?