Revert "Entity ID Serialisation (#56)"
This reverts commit 5ad8db2040fa39859a3846062c854ae9fbb802f3.

This was a bad design choice, we can do better. The issue here is that
if we were to try and load two scene files and both of them contained an
entity that used the same slot, we would crash. We want to be able to do
this in the future when we implement prefabs.

The design flaw here is making the index of the entity part of the
unique identifier for the entity. This stems from the original goal of
having all entities be represented by a uint32. What we should instead
do is decouple these; have a integer representing the index of the
entity, and additionally have a unique identier. We sort of do this now
anyway with the index/version making up the ID. We should replace the
version with a uuid and use *this* as the unique identifier, and have it
be completely unrelated to the index.

With this new idea, the uuid can be serialised, and when reloading an
entity, they will have the same identifier and potentially a different
index, which will stop clashes when loading multiple files. We can also
remove the restriction we placed on ourselves in the last commit of
requiring loading files before adding systems. With the new way we will
no longer need to clear all systems when loading and we will no longer
need to expose the internal structures of the ecs::Registry.

This will let us remove the SlotInfo functions which were a code smell
anyway.